backend/requirements.txt
flask
python-dotenv
requests
flask-cors
pyjwt
